Remove info_window changes
